% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CI_overlap.R
\name{check_CI_overlap}
\alias{check_CI_overlap}
\title{Check for Overlap of Confidence Intervals}
\usage{
check_CI_overlap(df, show_overlap = TRUE)
}
\arguments{
\item{df}{A data frame containing the columns 'tau', 'lower', and 'upper'.
This data frame can be the output from the `calculate_quantile_regression_CI` function.}

\item{show_overlap}{A logical value. If TRUE, shows pairs of taus with overlapping confidence intervals; 
if FALSE, shows pairs with non-overlapping intervals.}
}
\value{
A data frame with each pair of taus and an indication of whether their confidence intervals overlap. 
Returns a message if no intervals overlap (when show_overlap is TRUE) or if all intervals overlap (when show_overlap is FALSE).
}
\description{
This function checks for the overlap of confidence intervals for each pair of taus in a quantile regression model. 
It can be particularly useful for analyzing the output of the `calculate_quantile_regression_CI` function, 
which computes bootstrap confidence intervals for quantile regression coefficients.
}
\examples{
# Example of using calculate_quantile_regression_CI and check_CI_overlap together
# taus <- c(0.25, 0.5, 0.75)
# data <- data.frame(y = rnorm(100), x = rnorm(100))
# ci_results <- calculate_quantile_regression_CI(taus, independent_var, dependent_var, data, R = 100, ci_level = 0.95)
# check_results <- check_CI_overlap(ci_results, show_overlap = TRUE)

}
